+++ lib/great_list_view.dart
@@ -1,5 +1,8 @@
 library great_list_view;
 
-export 'src/core.dart';
+export 'src/core/core.dart';
 export 'src/tree_list_adapter.dart';
 export 'src/other_widgets.dart';
+export 'src/delegates.dart';
+export 'src/widgets.dart';
+export 'src/dispatcher.dart';
