+++ lib/src/core/child_manager.dart
@@ -6,7 +6,7 @@ part of 'core.dart';
 /// the children of subclasses of [RenderSliverMultiBoxAdaptor].
 ///
 /// This class takes the implementation of the standard [SliverMultiBoxAdaptorElement] class
-/// and provides support for animations and reordering feature.
+/// and adds support for animations and reordering feature.
 class AnimatedSliverMultiBoxAdaptorElement extends RenderObjectElement
     implements
         RenderSliverBoxChildManager,
@@ -18,7 +18,7 @@ class AnimatedSliverMultiBoxAdaptorElement extends RenderObjectElement
       AnimatedSliverMultiBoxAdaptorWidget widget)
       : super(widget);
 
-  // support for the method didChangeDependencies like in State.didChangeDependencies.
+  /// support for the method didChangeDependencies like in [State.didChangeDependencies].
   bool _didChangeDependencies = false;
 
   @override
@@ -29,7 +29,7 @@ class AnimatedSliverMultiBoxAdaptorElement extends RenderObjectElement
   AnimatedRenderSliverMultiBoxAdaptor get renderObject =>
       super.renderObject as AnimatedRenderSliverMultiBoxAdaptor;
 
-  _IntervalList get intervalList => renderObject._intervalList;
+  _IntervalList get _intervalList => renderObject._intervalList;
 
   /// This method has been overridden to give the render object the `didChangeDependencies` method
   /// and to link this list view to its controller.
@@ -63,10 +63,75 @@ class AnimatedSliverMultiBoxAdaptorElement extends RenderObjectElement
       oldWidget.listController._unsetInterface(this);
       newWidget.listController._setInterface(this);
     }
-    if (intervalList.hasPendingUpdates) {
+    if (_intervalList.hasPendingUpdates) {
       performRebuild();
     }
   }
+
+  // This function takes the old index of a child and calculates the new one by considering
+  // all the pending updates.
+  // This method can return `null` if the child no longer exists (for example,
+  // all elements of a remove interval disappear when it becomes a resizing interval).
+  ReindexResult oldIndexToNewIndex(List<_Update> updates, int index) {
+    var needsRebuild = false;
+    var clearLayoutOffset = false;
+    var reorderDrop = false;
+    var reorderPick = false;
+    var discardElement = false;
+    _PopUpList? popUpList;
+    for (final upd in updates) {
+      assert(upd.flags.checkIntegrity());
+      if (index >= upd.index && index < upd.index + upd.oldBuildCount) {
+        if (index >= upd.index + upd.newBuildCount) {
+          // is upd.newBuildCount < upd.oldBuildCount
+          return ReindexResult(null, needsRebuild, discardElement,
+              clearLayoutOffset, reorderPick, reorderDrop, popUpList);
+        }
+        if (upd.flags.hasClearLayoutOffset) {
+          if (index != upd.index || !upd.flags.hasKeepFirstLayoutOffset) {
+            clearLayoutOffset = true;
+          }
+        }
+        if (upd.flags.hasReorderDrop) {
+          if (reorderPick) {
+            reorderPick = false;
+            popUpList = null;
+          } else {
+            reorderDrop = true;
+            popUpList = upd.popUpList;
+          }
+        } else if (upd.flags.hasReorderPick) {
+          reorderPick = true;
+          popUpList = upd.popUpList;
+        }
+        needsRebuild = true;
+        if (upd.flags.hasDiscardElement) {
+          discardElement = true;
+        }
+      }
+      if (index >= upd.index + upd.oldBuildCount) index += upd.skipCount;
+    }
+    return ReindexResult(index, needsRebuild, discardElement, clearLayoutOffset,
+        reorderPick, reorderDrop, popUpList);
+  }
+
+  // String qqq() {
+  //   var s = "{";
+  //   _childElements.forEach((key, value) {
+  //     s += "$key: #${identityHashCode(value)}, ";
+  //   });
+  //   s += "}\n{";
+  //   var child = renderObject.firstChild;
+  //   while (child != null) {
+  //     final e = _childElements.values
+  //         .singleWhereOrNull((e) => e?.renderObject == child);
+  //     s +=
+  //         "${renderObject._parentDataOf(child)?.index}: #${e != null ? identityHashCode(e) : "NULL"}, ";
+  //     child = renderObject.childAfter(child);
+  //   }
+  //   s += "}";
+  //   return s;
+  // }
 
   /// Inspired by [SliverMultiBoxAdaptorElement.performRebuild].
   ///
@@ -74,7 +139,7 @@ class AnimatedSliverMultiBoxAdaptorElement extends RenderObjectElement
   /// have been added, old intervals have changed state, reordering has started, and so on).
   ///
   /// It considers all pending build updates to move each old last rendered child
-  /// to the correct new position.
+  /// to its new position.
   ///
   /// It also calls the [AnimatedRenderSliverList.didChangeDependencies] method if a dependency
   /// has been changed.
@@ -88,137 +153,113 @@ class AnimatedSliverMultiBoxAdaptorElement extends RenderObjectElement
     super.performRebuild();
 
     _currentBeforeChild = null;
-    var childrenUpdated = false;
+    // var childrenUpdated = false;
     assert(_currentlyUpdatingChildIndex == null);
     try {
       final newChildren = SplayTreeMap<int, Element?>();
-      final rebuildChildren = HashSet<int>();
-
-      for (final index in _childElements.keys.toList()) {
-        final childParentData = _childElements[index]?.renderObject?.parentData
-            as SliverMultiBoxAdaptorParentData?;
-
-        // This function takes the old index of a child and calculates the new one by considering
-        // all the pending updates.
-        // This method can return `null` if the child no longer exists (for example,
-        // all elements of a remove interval disappear when it becomes a resizing interval).
-        // Can set `needsRebuild` to `true` to mark the child as to be rebuilt.
-        // Can set `unbind` to `true` to invalidate its current layout offset (it is used by new
-        // resizing intervals in order to forget the layout offset of the previously built child,
-        // to avoid annoying scroll jumps when these intervals are not accurately measured).
-        var needsRebuild = false;
-        var unbind = false;
-        int? _oldIndexToNewIndex(int index) {
-          for (final upd in intervalList.updates) {
-            if (index >= upd.index && index < upd.index + upd.oldBuildCount) {
-              switch (upd.mode) {
-                case _UpdateMode.UNBIND:
-                  unbind |= unbind;
-                  return null;
-                case _UpdateMode.REPLACE:
-                  return null;
-                case _UpdateMode.REBUILD:
-                  if (index - upd.index >= upd.oldBuildCount + upd.skipCount) {
-                    return null;
-                  }
-                  needsRebuild = true;
-                  break;
-              }
-            }
-            if (index >= upd.index + upd.oldBuildCount) index += upd.skipCount;
-          }
-          return index;
-        }
-
-        var newIndex = _oldIndexToNewIndex(index);
-
-        if (childParentData != null && (newIndex == null || unbind)) {
-          childParentData.layoutOffset = null;
-        }
-
-        if (newIndex == null) {
-          newChildren.putIfAbsent(index, () => null);
-        } else {
-          if (needsRebuild) rebuildChildren.add(newIndex);
-
-          if (newIndex != index) {
-            newChildren[newIndex] = _childElements[index];
-            newChildren.putIfAbsent(index, () => null);
-            _childElements.remove(index);
-          } else {
-            newChildren.putIfAbsent(index, () => _childElements[index]);
-          }
-        }
-      }
-
-      void processElement(int index) {
-        _currentlyUpdatingChildIndex = index;
-        if (_childElements[index] != null &&
-            _childElements[index] != newChildren[index]) {
-          // This index has an old child that isn't used anywhere and should be deactivated.
-          _childElements[index] =
-              updateChild(_childElements[index], null, index);
-          childrenUpdated = true;
-        }
-        var oldChild = newChildren[index];
-        late final Element? newChild;
-        if (!rebuildChildren.contains(index) && oldChild != null) {
-          if ((oldChild.slot as int) != index) {
-            updateSlotForChild(oldChild, index);
-          }
-          newChild = oldChild;
-        } else {
-          newChild = updateChild(oldChild, _build(index), index);
-        }
-
-        if (newChild != null) {
-          childrenUpdated =
-              childrenUpdated || _childElements[index] != newChild;
-          _childElements[index] = newChild;
-          final parentData = newChild.renderObject!.parentData!
-              as SliverMultiBoxAdaptorParentData;
-          if (!parentData.keptAlive) {
-            _currentBeforeChild = newChild.renderObject as RenderBox?;
-          }
-        } else {
-          childrenUpdated = true;
-          _childElements.remove(index);
-        }
-      }
+      // final rebuildChildren = HashSet<int>();
 
       renderObject.debugChildIntegrityEnabled =
           false; // Moving children will temporary violate the integrity.
-      newChildren.keys.forEach(processElement);
+
+      for (final popUpList in _intervalList.popUpLists) {
+        final r = oldIndexToNewIndex(popUpList.updates, 0);
+        if (r.needsRebuild) {
+          if (popUpList is _ReorderPopUpList) {
+            _currentlyUpdatingChildIndex = null;
+            popUpList.element = updateChild(
+                popUpList.element, _build(0, popUpList: popUpList), null);
+          }
+        }
+        if (popUpList.interval == null) {
+          _intervalList.popUpLists.remove(popUpList);
+        }
+      }
+      for (int? index in _childElements.keys.toList()) {
+        Element? element = _childElements[index]!;
+        var childParentData = _parentDataOf(element);
+
+        final r = oldIndexToNewIndex(_intervalList.updates, index!);
+
+        if (r.reorderPick) {
+          assert(r.needsRebuild);
+          final rl = r.popUpList as _ReorderPopUpList;
+          rl.element = updateChild(element, _build(0, popUpList: rl), index)!;
+          element = null;
+          final renderBox = rl.element!.renderObject as RenderBox;
+          renderObject.remove(renderBox);
+          renderObject.adoptChild(renderBox);
+        }
+
+        if (r.reorderDrop) {
+          final layoutOffset = renderObject
+              .childScrollOffset(element!.renderObject as RenderBox);
+          _currentlyUpdatingChildIndex = index;
+          updateChild(element, null, index);
+          final rl = r.popUpList as _ReorderPopUpList;
+          element = rl.element!;
+          final renderBox = element.renderObject as RenderBox;
+          renderObject.dropChild(renderBox);
+          renderObject.insert(renderBox, after: _currentBeforeChild);
+          (renderBox.parentData! as SliverMultiBoxAdaptorParentData)
+              .layoutOffset = layoutOffset;
+          rl.element = null;
+        }
+
+        if (r.newIndex == null) {
+          updateChild(element, null, index);
+        } else {
+          if (childParentData != null && r.clearLayoutOffset) {
+            childParentData.layoutOffset = null;
+          }
+
+          _currentlyUpdatingChildIndex = r.newIndex;
+
+          if (r.needsRebuild) {
+            if (r.discardElement && element != null) {
+              element = updateChild(element, null, r.newIndex);
+            }
+            element = updateChild(element, _build(r.newIndex!), r.newIndex)!;
+          } else {
+            if (index != r.newIndex) {
+              // only update slot
+              updateChild(element, element!.widget, r.newIndex);
+            }
+          }
+
+          childParentData = _parentDataOf(element);
+          newChildren[r.newIndex!] = element!;
+          if (!childParentData!.keptAlive) {
+            _currentBeforeChild = element.renderObject as RenderBox?;
+          }
+        }
+      }
+
+      _childElements = newChildren;
     } finally {
       _currentlyUpdatingChildIndex = null;
+
       renderObject.debugChildIntegrityEnabled = true;
-      intervalList.updates.clear();
+      _intervalList.updates.clear();
     }
   }
 
-  Widget? _build(int index, [bool measureOnly = false]) {
-    final count = intervalList.buildItemCount;
-    if (index < 0 || index >= count) return null;
-    return intervalList.build(this, index, measureOnly);
-  }
-
-  // It creates an disposable off-list child, building the specified widget,
-  // and executes the callback. Eventually, the child is disposed.
-  void disposableChild(Widget widget, void Function(RenderBox) callback) {
-    owner!.lockState(() {
-      _currentlyUpdatingChildIndex = null;
-      Element? _measuringOffListChild;
-      _measuringOffListChild =
-          updateChild(_measuringOffListChild, widget, null);
-      assert(_measuringOffListChild != null);
-      callback(_measuringOffListChild!.renderObject! as RenderBox);
-      _measuringOffListChild = updateChild(_measuringOffListChild, null, null);
-      assert(_measuringOffListChild == null);
-    });
+  Widget? _build(int index, {bool measureOnly = false, _PopUpList? popUpList}) {
+    if (popUpList == null) {
+      final count = _intervalList.buildItemCount;
+      if (index < 0 || index >= count) return null;
+      return _intervalList.build(this, index, measureOnly);
+    } else {
+      final count = popUpList.interval?.popUpBuildCount ?? 0;
+      if (index < 0 || index >= count) return null;
+      final listIndex = _intervalList.listItemIndexOf(popUpList.interval!);
+      return popUpList.interval!
+          .buildPopUpWidget(this, index, listIndex, measureOnly);
+    }
   }
 
   @override
-  int get childCount => intervalList.buildItemCount;
+  int get childCount => _intervalList.buildItemCount;
 
   /// Copied from [SliverMultiBoxAdaptorElement.estimateMaxScrollOffset].
   /// The [AnimatedSliverChildBuilderDelegate.estimateMaxScrollOffset] method call
@@ -240,12 +281,11 @@ class AnimatedSliverMultiBoxAdaptorElement extends RenderObjectElement
   //
 
   int? _currentlyUpdatingChildIndex;
-  final SplayTreeMap<int, Element?> _childElements =
-      SplayTreeMap<int, Element?>();
+  SplayTreeMap<int, Element?> _childElements = SplayTreeMap<int, Element?>();
   RenderBox? _currentBeforeChild;
 
   /// Copied from [SliverMultiBoxAdaptorElement.didAdoptChild].
-  /// An assertion was removed to allow adoption of off-list elements.
+  /// An assertion has been removed to allow adoption of off-list elements.
   @override
   void didAdoptChild(final RenderBox child) {
     //""""""""""""""""""""""""""""""""""""""""""""""""""""""
@@ -361,7 +401,7 @@ class AnimatedSliverMultiBoxAdaptorElement extends RenderObjectElement
   }
 
   /// Copied from [SliverMultiBoxAdaptorElement.setDidUnderflow].
-  /// The didUnderflow variabile is no longer needed.
+  /// The `didUnderflow` variabile is no longer needed.
   @override
   void setDidUnderflow(final bool value) {
     // _didUnderflow = value;
@@ -376,8 +416,8 @@ class AnimatedSliverMultiBoxAdaptorElement extends RenderObjectElement
       renderObject.dropChild(child);
       return;
     }
+    // assert(_currentlyUpdatingChildIndex != null);
     //""""""""""""""""""""""""""""""""""""""""""""""""""""""
-    assert(_currentlyUpdatingChildIndex != null);
     renderObject.remove(child as RenderBox);
   }
 
@@ -401,8 +441,8 @@ class AnimatedSliverMultiBoxAdaptorElement extends RenderObjectElement
     assert(!_childElements.values.any((Element? child) => child == null));
     _childElements.values.cast<Element>().toList().forEach(visitor);
     //""""""""""""""""""""""""""""""""""""""""""""""""""""""
-    if (_reorderDraggedElement != null) {
-      visitor.call(_reorderDraggedElement!);
+    for (final popUpList in _intervalList.popUpLists) {
+      popUpList.elements.forEach((e) => visitor.call(e));
     }
     //""""""""""""""""""""""""""""""""""""""""""""""""""""""
   }
@@ -411,8 +451,7 @@ class AnimatedSliverMultiBoxAdaptorElement extends RenderObjectElement
   @override
   void debugVisitOnstageChildren(final ElementVisitor visitor) {
     _childElements.values.cast<Element>().where((Element child) {
-      final parentData =
-          child.renderObject!.parentData! as SliverMultiBoxAdaptorParentData;
+      final parentData = _parentDataOf(child)!;
       late double itemExtent;
       switch (renderObject.constraints.axis) {
         case Axis.horizontal:
@@ -439,18 +478,20 @@ class AnimatedSliverMultiBoxAdaptorElement extends RenderObjectElement
   }
 
   /// Copied from [SliverMultiBoxAdaptorElement.didFinishLayout].
-  /// The call of the didFinishLayout callback has been removed.
   @override
   void didFinishLayout() {
     assert(debugAssertChildListLocked());
-    //""""""""""""""""""""""""""""""""""""""""""""""""""""""
-    // final firstIndex = _childElements.firstKey() ?? 0;
-    // final lastIndex = _childElements.lastKey() ?? 0;
-    // widget.delegate.didFinishLayout(firstIndex, lastIndex);
-    //""""""""""""""""""""""""""""""""""""""""""""""""""""""
+    final firstIndex = _childElements.firstKey() ?? 0;
+    final lastIndex = _childElements.lastKey() ?? 0;
+    widget.delegate.didFinishLayout(firstIndex, lastIndex);
   }
 
-  int? _findBuildIndexFromContext(BuildContext context) {
+  //
+  // Support methods
+  //
+
+  /// Searchs and returns the list index of specified item by its [context], if it can be found.
+  int? buildIndexFromContext(BuildContext context) {
     if (context is! Element) return null;
     int? slot;
     if (context.slot is! int) {
@@ -480,133 +521,16 @@ class AnimatedSliverMultiBoxAdaptorElement extends RenderObjectElement
     return slot;
   }
 
-// *************************
-//  Reorder Feature Support
-// *************************
-
-  Element? _reorderDraggedElement;
-
-  void _pickUpDraggedItem(int buildIndex, int listIndex,
-      RenderBox? Function(RenderBox, double) callback) {
-    assert(_reorderDraggedElement == null);
-    owner!.buildScope(this, () {
-      // pop out the dragged child from the list
-      _reorderDraggedElement = _childElements[buildIndex];
-
-      // measure it
-      final itemSize = renderObject.measureItem(buildWidget(
-          this,
-          widget.delegate.builder,
-          listIndex,
-          AnimatedWidgetBuilderData(kAlwaysCompleteAnimation,
-              measuring: true, dragging: true)));
-
-      // delegate the work back to the render object
-      _currentBeforeChild =
-          callback(_reorderDraggedElement!.renderObject as RenderBox, itemSize);
-
-      // notify IntervalList that the reorder has been started:
-      // a _ReorderOpeningInterval is returned
-      final openInterval =
-          intervalList.notifyStartReorder(buildIndex, itemSize);
-
-      // replace the in-list dragged item with a new resizing interval widget
-      _currentlyUpdatingChildIndex = buildIndex;
-      final newElement = updateChild(
-          null, openInterval.buildWidget(this, 0, 0, false), buildIndex);
-      _childElements[buildIndex] = newElement;
+  /// It creates an disposable off-list element built with the specified [widget],
+  /// and executes the [callback]. Eventually, the child is disposed.
+  void _disposableElement(Widget widget, void Function(RenderBox) callback) {
+    owner!.lockState(() {
       _currentlyUpdatingChildIndex = null;
-
-      // rebuild dragged child with dragging attribute set
-      _reorderDraggedElement = updateChild(
-          _reorderDraggedElement,
-          buildWidget(
-              this,
-              widget.delegate.builder,
-              listIndex,
-              AnimatedWidgetBuilderData(kAlwaysCompleteAnimation,
-                  dragging: true, slot: renderObject._slot)),
-          buildIndex);
-    });
-  }
-
-  void _dropDraggedItem(int listItemIndex,
-      RenderBox? Function(RenderBox? currentRenderBox) callback) {
-    owner!.buildScope(this, () {
-      var draggedElement = _reorderDraggedElement!;
-
-      // notify IntervalList that the reorder has been stopped:
-      // the buildIndex where the dragged item will be dropped is returned
-      final buildIndex = intervalList.notifyStopReorder();
-
-      // retrieve the resizing interval child that will be replaced with the dragged item
-      final currentElement = _childElements[buildIndex];
-
-      // delegate the work back to the render object
-      final currentRenderBox = currentElement?.renderObject as RenderBox?;
-      _currentlyUpdatingChildIndex = buildIndex;
-      final previousRenderBox = callback(currentRenderBox);
-
-      if (currentRenderBox != null) {
-        // the drop zone is still visibile, the off-list dragged item must be rebuilt
-        // with the dragging attribute unset and put back in the list
-        _currentBeforeChild = previousRenderBox;
-        draggedElement = updateChild(
-            draggedElement,
-            buildWidget(
-                this,
-                widget.delegate.builder,
-                listItemIndex,
-                AnimatedWidgetBuilderData(kAlwaysCompleteAnimation,
-                    slot: renderObject._slot)),
-            buildIndex)!;
-        _childElements[buildIndex] = draggedElement;
-      } else {
-        // the drop zone is no longer visibile, the off-list dragged item can be disposed
-        _currentlyUpdatingChildIndex = null;
-        updateChild(draggedElement, null, null);
-      }
-
-      _reorderDraggedElement = null;
-      _currentlyUpdatingChildIndex = null;
-    });
-  }
-
-  void _disposeDraggedItem() {
-    assert(_reorderDraggedElement != null);
-    owner!.buildScope(this, () {
-      var draggedElement = _reorderDraggedElement!;
-      _currentlyUpdatingChildIndex = null;
-      _reorderDraggedElement = null;
-      updateChild(draggedElement, null, null);
-    });
-  }
-
-  void _rebuildDraggedItem(int itemIndex, double Function(Widget) callback) {
-    assert(_reorderDraggedElement != null);
-    owner!.buildScope(this, () {
-      final measuredWidget = buildWidget(
-          this,
-          widget.delegate.builder,
-          itemIndex,
-          AnimatedWidgetBuilderData(kAlwaysCompleteAnimation,
-              measuring: true, dragging: true, slot: renderObject._slot));
-      final newWidget = buildWidget(
-          this,
-          widget.delegate.builder,
-          itemIndex,
-          AnimatedWidgetBuilderData(kAlwaysCompleteAnimation,
-              dragging: true, slot: renderObject._slot));
-
-      // _currentlyUpdatingChildIndex = null;
-      _currentBeforeChild = null;
-
-      _reorderDraggedElement =
-          updateChild(_reorderDraggedElement, newWidget, null);
-
-      final newItemSize = callback(measuredWidget);
-
-      intervalList.reorderChangeOpeningIntervalSize(newItemSize);
+      var _measuringOffListChild = updateChild(null, widget, null);
+      assert(_measuringOffListChild != null);
+      callback(_measuringOffListChild!.renderObject! as RenderBox);
+      _measuringOffListChild = updateChild(_measuringOffListChild, null, null);
+      assert(_measuringOffListChild == null);
     });
   }
 
@@ -621,44 +545,39 @@ class AnimatedSliverMultiBoxAdaptorElement extends RenderObjectElement
   bool get isHorizontal => renderObject.constraints.axis == Axis.horizontal;
 
   @override
-  Widget buildWidget(BuildContext context, AnimatedWidgetBuilder builder,
-      int index, AnimatedWidgetBuilderData data) {
+  Widget buildWidget(AnimatedWidgetBuilder builder, int index,
+      AnimatedWidgetBuilderData data) {
     Widget child;
     try {
-      child = builder.call(context, index, data);
+      child = builder.call(this, index, data);
     } catch (exception, stackTrace) {
       child = _createErrorWidget(exception, stackTrace);
     }
-    return widget.delegate.wrapWidget(context, child, data);
+    return widget.delegate.wrapWidget(this, child, data);
   }
 
-  // It notifies that the resizing interval has changed its size by the delta amount.
+  /// Notifies that the resizing interval has changed its size by the delta amount.
   @override
-  void resizingIntervalUpdated(_ResizingInterval interval, double delta) {
+  void resizingIntervalUpdated(_AnimatedSpaceInterval interval, double delta) {
     renderObject._resizingIntervalUpdated(interval, delta);
   }
 
   /// Measure the size of a bunch of off-list children up to [count] elements.
   /// You have to provide a [builder] to build the `i`-th widget.
-  /// The calculation is asynchronous and can be cancelled.
+  /// The calculation is asynchronous and can be [cancelled].
   @override
   Future<_Measure> measureItems(
-      _Cancelled cancelled, int count, IndexedWidgetBuilder builder) async {
+      _Cancelled? cancelled, int count, IndexedWidgetBuilder builder) async {
     return await renderObject.measureItems(cancelled, count, builder);
   }
 
   @override
-  void draggedItemHasBeenRemoved() {
-    renderObject._reorderDraggedItemHasBeenRemoved();
-  }
-
-  @override
-  void draggedItemHasChanged() {
-    renderObject._reorderDraggedItemHasChanged();
+  double measureItem(Widget widget) {
+    return renderObject.measureItem(widget);
   }
 
   //
-  // Implementation of _ControllerListeners
+  // Implementation of _ControllerInterface
   //
 
   int _batch = 0;
@@ -670,7 +589,7 @@ class AnimatedSliverMultiBoxAdaptorElement extends RenderObjectElement
     callback();
     _batch--;
     assert(_batch >= 0);
-    if (_batch == 0) intervalList.coordinate();
+    if (_batch == 0) _intervalList.coordinate();
   }
 
   /// See [AnimatedListController.notifyInsertedRange].
@@ -691,29 +610,25 @@ class AnimatedSliverMultiBoxAdaptorElement extends RenderObjectElement
       _notifyReplacedRange(
           from, removeCount, insertCount, removedItemBuilder, priority);
 
+  void _notifyReplacedRange(int from, int removeCount, int insertCount,
+      final AnimatedWidgetBuilder? removedItemBuilder, int priority) {
+    _intervalList.notifyReplacedRange(
+        from, removeCount, insertCount, removedItemBuilder, priority);
+    if (_batch == 0) _intervalList.coordinate();
+  }
+
   /// See [AnimatedListController.notifyChangedRange].
   @override
   void notifyChangedRange(int from, int count,
-          final AnimatedWidgetBuilder changedItemBuilder, int priority) =>
-      _notifyChangedRange(from, count, changedItemBuilder, priority);
-
-  void _notifyReplacedRange(int from, int removeCount, final int insertCount,
-      final AnimatedWidgetBuilder? removedItemBuilder, int priority) {
-    intervalList.notifyRangeReplaced(
-        from, removeCount, insertCount, removedItemBuilder, priority);
-    if (_batch == 0) intervalList.coordinate();
-  }
-
-  void _notifyChangedRange(int from, int count,
       final AnimatedWidgetBuilder changedItemBuilder, int priority) {
-    intervalList.notifyRangeChanged(from, count, changedItemBuilder, priority);
-    if (_batch == 0) intervalList.coordinate();
+    _intervalList.notifyChangedRange(from, count, changedItemBuilder, priority);
+    if (_batch == 0) _intervalList.coordinate();
   }
 
   /// See [AnimatedListController.notifyStartReorder].
   @override
-  void notifyStartReorder(BuildContext context, double dx, double dy) {
-    renderObject._reorderStart(context, dx, dy);
+  bool notifyStartReorder(BuildContext context, double dx, double dy) {
+    return renderObject._reorderStart(context, dx, dy);
   }
 
   /// See [AnimatedListController.notifyUpdateReorder].
@@ -728,52 +643,32 @@ class AnimatedSliverMultiBoxAdaptorElement extends RenderObjectElement
     renderObject._reorderStop(cancel);
   }
 
+  /// See [AnimatedListController.computeItemBox].
   @override
   Rect? computeItemBox(int index, bool absolute) {
-    return renderObject._computeItemBox(index, absolute);
+    return renderObject._computeItemBox(index, absolute, false);
   }
 
+  /// See [AnimatedListController.getItemVisibleSize].
   @override
-  int? listToActualItemIndex(int index) {
-    if (index < 0 || index >= intervalList.listItemCount) {
-      return null;
-    }
-    final interval = intervalList.intervalAtListIndex(index);
-    if (interval.interval is _InListItemInterval ||
-        interval.interval is _ReadyToChangingInterval) {
-      return interval.buildIndex + (index - interval.itemIndex);
-    }
-    return null;
-  }
+  PercentageSize? getItemVisibleSize(int index) {
+    final box = renderObject._computeItemBox(index, true, true);
+    if (box != null) {
+      final c = renderObject.constraints;
+      final v1 = c.scrollOffset +
+          c.precedingScrollExtent -
+          (c.viewportMainAxisExtent - c.remainingPaintExtent);
+      final v2 = v1 + c.viewportMainAxisExtent;
 
-  @override
-  int? actualToListItemIndex(int index) {
-    if (index < 0 || index >= intervalList.buildItemCount) {
-      return null;
+      final r1 = (isHorizontal ? box.left : box.top).clamp(v1, v2);
+      final r2 = (isHorizontal ? box.right : box.bottom).clamp(v1, v2);
+      return PercentageSize(
+          math.max(0.0, r2 - r1), (isHorizontal ? box.width : box.height));
     }
-    final interval = intervalList.intervalAtBuildIndex(index);
-    if (interval.interval is _InListItemInterval ||
-        interval.interval is _ReadyToChangingInterval) {
-      return interval.itemIndex + (index - interval.buildIndex);
-    }
-    return null;
   }
 }
 
-abstract class _ListIntervalInterface {
-  AnimatedSliverChildDelegate get delegate;
-  bool get isHorizontal;
-  Widget buildWidget(BuildContext context, AnimatedWidgetBuilder builder,
-      int index, AnimatedWidgetBuilderData data);
-  void resizingIntervalUpdated(_ResizingInterval interval, double delta);
-  Future<_Measure> measureItems(
-      _Cancelled cancelled, int count, IndexedWidgetBuilder builder);
-  void draggedItemHasBeenRemoved() {}
-  void draggedItemHasChanged() {}
-  void markNeedsBuild();
-}
-
-// Return a Widget for the given Exception (copied from the standard package).
+// Returns a Widget for the given Exception (copied from the standard package).
 Widget _createErrorWidget(Object exception, StackTrace stackTrace) {
   final details = FlutterErrorDetails(
     exception: exception,
@@ -783,4 +678,38 @@ Widget _createErrorWidget(Object exception, StackTrace stackTrace) {
   );
   FlutterError.reportError(details);
   return ErrorWidget.builder(details);
+}
+
+class ReindexResult {
+  /// The new index of the child.
+  final int? newIndex;
+
+  /// Marks the child as to be rebuilt.
+  final bool needsRebuild;
+
+  /// Marks the child as to be rebuilt as new, discarding its previous [Element].
+  final bool discardElement;
+
+  /// Marks to invalidate its current layout offset (especially used by new resizing intervals in order
+  /// to forget the layout offset of the previously built child, to avoid annoying scroll jumps when
+  /// these intervals are not accurately measured).
+  final bool clearLayoutOffset;
+
+  /// Marks the child to be picked up and its [Element] to be moved into the [popUpList].
+  final bool reorderPick;
+
+  /// Marks the child to be dropped and its [Element] to be moved from the [popUpList] into the main list.
+  final bool reorderDrop;
+
+  /// The [_PopUpList] referred to [reorderPick] or [reorderDrop].
+  final _PopUpList? popUpList;
+
+  const ReindexResult(
+      this.newIndex,
+      this.needsRebuild,
+      this.discardElement,
+      this.clearLayoutOffset,
+      this.reorderPick,
+      this.reorderDrop,
+      this.popUpList);
 }
