+++ lib/src/core/sliver_list.dart
@@ -9,8 +9,7 @@ bool fixScrollableRepainting = true;
 /// This mixin is used by [AnimatedRenderSliverList] and [AnimatedRenderSliverFixedExtentList].
 mixin AnimatedRenderSliverMultiBoxAdaptor
     implements RenderSliverMultiBoxAdaptor {
-  bool _inPaint = false;
-  double _resizeAmountCorrection = 0.0;
+  double _resizeCorrectionAmount = 0.0;
 
   late _IntervalList _intervalList;
 
@@ -21,18 +20,19 @@ mixin AnimatedRenderSliverMultiBoxAdaptor
     _intervalList = _IntervalList(childManager);
   }
 
+  @override
   void dispose() {
     _intervalList.dispose();
   }
 
-  /// It returns the sum of the sizes of a bunch of items, builded by the specified builder.
+  /// Returns the sum of the sizes of a bunch of items, built using the specified builder.
   Future<_Measure> measureItems(
-      _Cancelled cancelled, int count, IndexedWidgetBuilder builder);
+      _Cancelled? cancelled, int count, IndexedWidgetBuilder builder);
 
-  /// It returns the size of a single widget.
+  /// Returns the size of a single widget.
   double measureItem(Widget widget, [BoxConstraints? childConstraints]);
 
-  /// It calculates an estimate of the maximum scroll offset.
+  /// Calculates an estimate of the maximum scroll offset.
   double? extrapolateMaxScrollOffset(
     final int firstIndex,
     final int lastIndex,
@@ -44,30 +44,38 @@ mixin AnimatedRenderSliverMultiBoxAdaptor
   /// The specified interval has been resized by a delta amount.
   /// If this interval is above the first item layouted in the list view, this amount will be
   /// added up to a cumulative variabile.
-  void _resizingIntervalUpdated(_ResizingInterval interval, double delta) {
+  void _resizingIntervalUpdated(_AnimatedSpaceInterval interval, double delta) {
     assert(delta != 0.0);
     final firstChild = firstChildWithLayout;
     if (firstChild == null) return;
-    if (indexOf(firstChild) <= _intervalList.buildItemIndexOf(interval)) return;
-    _resizeAmountCorrection += delta;
+    if (_intervalList.buildItemIndexOf(interval) < indexOf(firstChild)) {
+      _resizeCorrectionAmount += delta;
+    }
     markNeedsLayout();
   }
 
-  // Adjusts the layout offset of the first displayed item by the correction amount
-  // calcoluted as a result of resizing the intervals above it.
-  void _adjustLayout() {
-    var amount = _resizeAmountCorrection;
-    _resizeAmountCorrection = 0.0;
+  // Adjusts the layout offset of the first layouted item by the correction amount
+  // calculated as a result of resizing the intervals above it.
+  bool _adjustTopLayoutOffset() {
+    var amount = _resizeCorrectionAmount;
+    _resizeCorrectionAmount = 0.0;
 
     final firstChild = firstChildWithLayout;
     if (firstChild != null) {
-      final parentData = parentDataOf(firstChild)!;
+      final parentData = _parentDataOf(firstChild)!;
       final firstLayoutOffset = parentData.layoutOffset!;
       if (firstLayoutOffset + amount < 0.0) {
-        amount = -firstLayoutOffset; // bring back first offset to zero
+        amount = -firstLayoutOffset; // bring back the offset to zero
       }
-      parentData.layoutOffset = parentData.layoutOffset! + amount;
+      if (amount != 0.0) {
+        parentData.layoutOffset = parentData.layoutOffset! + amount;
+        if ( childManager.widget.delegate.holdScrollOffset ) {
+        geometry = SliverGeometry(scrollOffsetCorrection: amount);
+        return false;
+        }
+      }
     }
+    return true;
   }
 
   /// A `didChangeDependencies` method like in [State.didChangeDependencies],
@@ -76,7 +84,7 @@ mixin AnimatedRenderSliverMultiBoxAdaptor
     _intervalList.updateTickerMuted(context);
   }
 
-  // It returns the first displayed and layouted item.
+  // Returns the first displayed and layouted item.
   RenderBox? get firstChildWithLayout {
     var child = firstChild;
     while (child != null) {
@@ -96,11 +104,11 @@ mixin AnimatedRenderSliverMultiBoxAdaptor
     }
   }
 
-  SliverMultiBoxAdaptorParentData? parentDataOf(RenderBox child) =>
-      child.parentData as SliverMultiBoxAdaptorParentData?;
+  SliverMultiBoxAdaptorParentData? _parentDataOf(RenderBox? child) =>
+      child?.parentData as SliverMultiBoxAdaptorParentData?;
 
-  // It dispatches a fake change to the ScrollPosition to force the listeners
-  // (ie a Scrollbar) to refresh its state.
+  /// Dispatches a fake change to the [ScrollPosition] to force the listeners
+  /// (ie a [Scrollbar]) to refresh its state.
   void _notifyScrollable() {
     final scrollable = Scrollable.of(childManager);
     if (scrollable == null) return;
@@ -117,7 +125,9 @@ mixin AnimatedRenderSliverMultiBoxAdaptor
 
   bool _initScrollPosition = true;
 
-  void doPerformLayout(void Function() callback) {
+  void doPerformLayout(VoidCallback callback) {
+    if (!_adjustTopLayoutOffset()) return;
+
     if (_initScrollPosition) {
       _initScrollPosition = false;
       final callback = childManager.widget.delegate.initialScrollOffsetCallback;
@@ -135,15 +145,13 @@ mixin AnimatedRenderSliverMultiBoxAdaptor
       }
     }
 
-    _adjustLayout();
-
     callback();
 
     if ((geometry!.scrollOffsetCorrection ?? 0.0) != 0.0) {
       return;
     }
 
-    _reoderPerformLayout();
+    _reorderPerformLayout();
 
     if (fixScrollableRepainting) {
       WidgetsBinding.instance?.addPostFrameCallback((_) {
@@ -154,70 +162,86 @@ mixin AnimatedRenderSliverMultiBoxAdaptor
 
   RenderBox? _childAfter(RenderBox child) {
     assert(child.parent == this);
-    return parentDataOf(child)?.nextSibling;
+    return _parentDataOf(child)?.nextSibling;
+  }
+
+  Iterable<RenderBox> allChildren(RenderBox? firstChild) sync* {
+    var child = firstChild;
+    while (child != null) {
+      yield child;
+      child = _childAfter(child);
+    }
+    for (final popUpList in _intervalList.popUpLists) {
+      for (final element in popUpList.elements) {
+        yield element.renderObject as RenderBox;
+      }
+    }
   }
 
   @override
   RenderBox? childAfter(RenderBox child) {
-    if (_inPaint && _reorderDraggedRenderBox != null) {
-      // the dragged item is the last child
-      if (child == _reorderDraggedRenderBox) {
+    if (_popUpChildrenIterator != null) {
+      if (!_popUpChildrenIterator!.moveNext()) {
         return null;
+      } else {
+        return _popUpChildrenIterator!.current;
       }
-
-      var after = _childAfter(child);
-      after ??= _reorderDraggedRenderBox;
-
-      return after;
     }
     return _childAfter(child);
   }
 
+  Iterator<RenderBox>? _popUpChildrenIterator;
+
   void doPaint(
       PaintingContext context, Offset offset, void Function() callback) {
-    _inPaint = true;
+    _popUpChildrenIterator = allChildren(firstChild).skip(1).iterator;
     callback();
-    _inPaint = false;
+    _popUpChildrenIterator = null;
   }
 
   void doVisitChildren(RenderObjectVisitor visitor, void Function() callback) {
     callback();
-    if (_reorderDraggedRenderBox != null) {
-      visitor.call(_reorderDraggedRenderBox!);
+    for (final popUpList in _intervalList.popUpLists) {
+      for (final element in popUpList.elements) {
+        visitor.call(element.renderObject!);
+      }
     }
   }
 
-  // *************************
-  //  Reorder Feature Support
-  // *************************
+  void layoutWithScrollOffset(RenderBox renderBox, double layoutOffset) {
+    renderBox.layout(constraints.asBoxConstraints(), parentUsesSize: true);
+    final parentData = renderBox.parentData! as SliverMultiBoxAdaptorParentData;
+    parentData.layoutOffset = layoutOffset;
+  }
 
-  RenderBox? _reorderDraggedRenderBox;
-  var _reorderDraggedItemSize = 0.0;
+  //
+  // Reorder Feature Support
+  //
 
-  var _reorderMainAxisLocalOffset = 0.0;
-  var _reorderCrossAxisLocalOffset = 0.0;
+  _ReorderPopUpList? get _reorderPopUpList =>
+      _intervalList.popUpLists.singleWhereOrNull((e) => e is _ReorderPopUpList)
+          as _ReorderPopUpList?;
 
-  var _reorderOriginScrollOffset = 0.0;
+  RenderBox? get _reorderDraggedRenderBox =>
+      _reorderPopUpList?.element?.renderObject as RenderBox?;
 
-  var _reorderOriginMainAxisOffset = 0.0;
-  var _reorderOriginCrossAxisOffset = 0.0;
+  bool get isReordering =>
+      _intervalList.any((e) => e is _ReorderHolderNormalInterval);
 
-  var _reorderCurrentMainAxisOffset = 0.0;
-  var _reorderLastMainAxisOffset = 0.0;
+  double get _reorderDraggedItemSize => _reorderPopUpList!.itemSize;
 
-  Object? _slot;
+  final _reorderLayoutData = _ReorderLayoutData();
 
-  bool get isReordering => _reorderDraggedRenderBox != null;
-
-  // Initiazes a new reorder.
-  // The onReorderStartcallback will be invoked.
-  // Returns true if the reorder can be performed.
-  // The picked up item will be removed from the children list in order to be painted above all of them.
-  bool _reorderStart(BuildContext itemContext, double dx, double dy) {
+  bool _reorderStart(BuildContext context, double dx, double dy) {
     if (isReordering) return false;
 
-    var buildIndex = childManager._findBuildIndexFromContext(itemContext);
-    if (buildIndex == null) return false;
+    var oldBuildIndex = childManager.buildIndexFromContext(context);
+    if (oldBuildIndex == null) return false;
+
+    final r =
+        childManager.oldIndexToNewIndex(_intervalList.updates, oldBuildIndex);
+    if (r.newIndex == null || r.needsRebuild) return false;
+    final buildIndex = r.newIndex!;
 
     final info = _intervalList.intervalAtBuildIndex(buildIndex);
     if (info.interval is! _NormalInterval) return false;
@@ -228,72 +252,44 @@ mixin AnimatedRenderSliverMultiBoxAdaptor
     if (!(delegate.reorderModel?.onReorderStart.call(itemIndex, dx, dy) ??
         false)) return false;
 
-    switch (constraints.axis) {
-      case Axis.horizontal:
-        _reorderMainAxisLocalOffset = dx;
-        _reorderCrossAxisLocalOffset = dy;
-        break;
-      case Axis.vertical:
-        _reorderCrossAxisLocalOffset = dx;
-        _reorderMainAxisLocalOffset = dy;
-        break;
-    }
+    _reorderLayoutData.init(
+        constraints,
+        _parentDataOf(childManager._childElements[oldBuildIndex]!.renderObject
+                as RenderBox)!
+            .layoutOffset!,
+        dx,
+        dy);
 
-    childManager._pickUpDraggedItem(buildIndex, itemIndex,
-        (RenderBox draggedRenderBox, double itemSize) {
-      final oldChildParentData =
-          draggedRenderBox.parentData! as SliverMultiBoxAdaptorParentData;
+    final slot = delegate.reorderModel?.onReorderFeedback.call(itemIndex,
+        itemIndex, _reorderLayoutData.currentMainAxisOffset, 0.0, 0.0);
 
-      final previousRenderBox = childBefore(draggedRenderBox);
-
-      _reorderDraggedItemSize = itemSize;
-
-      remove(draggedRenderBox);
-      adoptChild(draggedRenderBox);
-
-      draggedRenderBox.layout(constraints.asBoxConstraints(),
-          parentUsesSize: true);
-      final newChildParentData =
-          draggedRenderBox.parentData! as SliverMultiBoxAdaptorParentData;
-      newChildParentData.layoutOffset = oldChildParentData.layoutOffset;
-
-      _reorderOriginMainAxisOffset =
-          newChildParentData.layoutOffset! - _reorderMainAxisLocalOffset;
-      _reorderDraggedRenderBox = draggedRenderBox;
-      _reorderOriginCrossAxisOffset = _reorderCrossAxisLocalOffset;
-
-      return previousRenderBox;
-    });
-
-    _slot = delegate.reorderModel?.onReorderFeedback
-        .call(itemIndex, itemIndex, _reorderCurrentMainAxisOffset, 0.0, 0.0);
-
-    _reorderOriginScrollOffset = constraints.scrollOffset;
-
-    _reorderCurrentMainAxisOffset =
-        _reorderOriginMainAxisOffset + _reorderMainAxisLocalOffset;
-    _reorderLastMainAxisOffset = _reorderOriginMainAxisOffset;
-
-    markNeedsLayout();
+    // notify IntervalList that the reorder has been started
+    _intervalList.notifyStartReorder(itemIndex, slot);
 
     return true;
   }
 
-  // It calls the onReorderFeedback callback.
-  // If the slot is changed, the dragged item will be rebuilded.
+  // The dragged item has been released, so the reordering can be completed.
+  void _reorderStop(bool cancel) {
+    if (!isReordering) return;
+
+    final pickIndex = _intervalList.reorderPickListIndex!;
+    final dropIndex = _intervalList.reorderDropListIndex(pickIndex);
+
+    final delegate = childManager.widget.delegate;
+
+    cancel = cancel ||
+        !(delegate.reorderModel?.onReorderComplete
+                .call(pickIndex, dropIndex, _reorderPopUpList?.slot) ??
+            false);
+
+    _intervalList.notifyStopReorder(cancel);
+  }
+
   void _reorderUpdate(double dx, double dy) {
     if (!isReordering) return;
 
-    switch (constraints.axis) {
-      case Axis.horizontal:
-        _reorderMainAxisLocalOffset = dx;
-        _reorderCrossAxisLocalOffset = dy;
-        break;
-      case Axis.vertical:
-        _reorderCrossAxisLocalOffset = dx;
-        _reorderMainAxisLocalOffset = dy;
-        break;
-    }
+    _reorderLayoutData.update(constraints, dx, dy);
 
     final delegate = childManager.widget.delegate;
     final pickIndex = _intervalList.reorderPickListIndex!;
@@ -301,47 +297,36 @@ mixin AnimatedRenderSliverMultiBoxAdaptor
     final newSlot = delegate.reorderModel?.onReorderFeedback.call(
         pickIndex,
         dropIndex,
-        _reorderCurrentMainAxisOffset,
-        _reorderCrossAxisLocalOffset - _reorderOriginCrossAxisOffset,
-        _reorderMainAxisLocalOffset);
-    if (_slot != newSlot) {
-      _slot = newSlot;
+        _reorderLayoutData.currentMainAxisOffset,
+        _reorderLayoutData.crossAxisDelta,
+        _reorderLayoutData.axisLocalOffset);
 
-      childManager.notifyChangedRange(
-          pickIndex,
-          1,
-          (context, index, data) => childManager.widget.delegate
-              .builder(context, _intervalList.reorderPickListIndex!, data),
-          0);
-    }
+    _reorderPopUpList!.updateSlot(newSlot);
 
     markNeedsLayout();
   }
 
   // Compute whether the dragged item has been moved up or down by comparing the new position with the old one.
   void _reorderComputeNewOffset() {
-    assert(isReordering);
-    final current = _reorderCurrentMainAxisOffset;
-    if (current != _reorderLastMainAxisOffset) {
-      if (current > _reorderLastMainAxisOffset) {
-        _reorderMoveDownwards();
-      } else {
-        _reorderMoveUpwards();
-      }
-      _reorderLastMainAxisOffset = current;
+    if (!isReordering) return;
+    final dir = _reorderLayoutData.updateDirection();
+    if (dir == 1) {
+      _reorderMoveDownwards();
+    } else if (dir == -1) {
+      _reorderMoveUpwards();
     }
   }
 
   // Calculates the new drop position of the item moved down.
   void _reorderMoveDownwards() {
     final draggedBottom =
-        _reorderCurrentMainAxisOffset + _reorderDraggedItemSize;
+        _reorderLayoutData.currentMainAxisOffset + _reorderDraggedItemSize;
 
     RenderBox? child;
     _NormalInterval? lastNormalInterval;
     var lastBuildIndex = 0;
     for (child = lastChild; child != null; child = childBefore(child)) {
-      final childParentData = parentDataOf(child)!;
+      final childParentData = _parentDataOf(child)!;
       final childBottom =
           childParentData.layoutOffset! + childSize(child) * 0.75;
 
@@ -362,13 +347,13 @@ mixin AnimatedRenderSliverMultiBoxAdaptor
 
   // Calculates the new drop position of the item moved up.
   void _reorderMoveUpwards() {
-    final draggedTop = _reorderCurrentMainAxisOffset;
+    final draggedTop = _reorderLayoutData.currentMainAxisOffset;
 
     RenderBox? child;
     _NormalInterval? lastNormalInterval;
     var lastBuildIndex = 0;
     for (child = firstChild; child != null; child = childAfter(child)) {
-      final childParentData = parentDataOf(child)!;
+      final childParentData = _parentDataOf(child)!;
       final childTop = childParentData.layoutOffset! + childSize(child) * 0.25;
 
       if (draggedTop < childTop) {
@@ -386,8 +371,9 @@ mixin AnimatedRenderSliverMultiBoxAdaptor
     _changeDropIndex(lastBuildIndex, lastNormalInterval, true);
   }
 
-  // Calls the onReorderMove callback to find out if the new drop position is allowed.
-  // If so, the intervals are updated accordingly.
+  /// Calls the [AnimatedListBaseReorderModel.onReorderMove] callback to find out if the
+  /// new drop position is allowed.
+  /// If so, the intervals are updated accordingly.
   void _changeDropIndex(
       int buildIndex, _NormalInterval interval, bool onTheRight) {
     final intervalListIndex = _intervalList.listItemIndexOf(interval);
@@ -408,88 +394,11 @@ mixin AnimatedRenderSliverMultiBoxAdaptor
         insertBuildIndex - intervalBuildIndex, _reorderDraggedItemSize);
   }
 
-  // The dragged item has been released, so the reodering will be completed.
-  void _reorderStop(bool cancel) {
-    if (!isReordering) return;
-
-    final pickIndex = _intervalList.reorderPickListIndex!;
-    var dropIndex = _intervalList.reorderDropListIndex(pickIndex);
-
-    final delegate = childManager.widget.delegate;
-    if (cancel ||
-        !(delegate.reorderModel?.onReorderComplete
-                .call(pickIndex, dropIndex, _slot) ??
-            false)) {
-      dropIndex = pickIndex;
-    }
-
-    _slot = null;
-
-    childManager._dropDraggedItem(dropIndex, (RenderBox? currentRenderBox) {
-      if (currentRenderBox == null) {
-        _reorderDraggedRenderBox = null;
-        return null;
-      }
-
-      final currentChildParentData =
-          currentRenderBox.parentData! as SliverMultiBoxAdaptorParentData;
-
-      dropChild(_reorderDraggedRenderBox!);
-
-      final previousRenderBox = childBefore(currentRenderBox);
-      remove(currentRenderBox);
-      insert(_reorderDraggedRenderBox!, after: previousRenderBox);
-
-      _reorderDraggedRenderBox!
-          .layout(constraints.asBoxConstraints(), parentUsesSize: true);
-      parentDataOf(_reorderDraggedRenderBox!)!.layoutOffset =
-          currentChildParentData.layoutOffset;
-
-      _reorderDraggedRenderBox = null;
-
-      return previousRenderBox;
-    });
-
-    markNeedsLayout();
-  }
-
-  // The dragged item has been removed, the reorder will be aborted.
-  void _reorderDraggedItemHasBeenRemoved() {
-    assert(isReordering);
-    _reorderDraggedRenderBox = null;
-    childManager._disposeDraggedItem();
-    markNeedsLayout();
-  }
-
-  // The dragged item has changed, so this item has be rebuilt and the open reorder interval
-  // will eventually be recreated in order to make room for the new item size.
-  void _reorderDraggedItemHasChanged() {
-    final itemIndex = _intervalList.reorderPickListIndex;
-    if (itemIndex == null) return;
-
-    final oldScrollOffset = childScrollOffset(_reorderDraggedRenderBox!);
-
-    childManager._rebuildDraggedItem(itemIndex, (Widget measuredWidget) {
-      final childConstraints = constraints.asBoxConstraints();
-
-      final newItemSize = measureItem(measuredWidget, childConstraints);
-
-      _reorderDraggedItemSize = newItemSize;
-
-      _reorderDraggedRenderBox!.layout(childConstraints, parentUsesSize: true);
-      final newChildParentData = _reorderDraggedRenderBox!.parentData!
-          as SliverMultiBoxAdaptorParentData;
-      newChildParentData.layoutOffset = oldScrollOffset;
-
-      return newItemSize;
-    });
-  }
-
   // Updates the layout offset of the dragged item.
   // Also schedules the calculation of the new drop position.
   // Finally, it considers wheter to trigger a scroling of the list because
   // the dragged item has been moved the the top or bottom of it.
-  void _reoderPerformLayout() {
+  void _reorderPerformLayout() {
     if (!isReordering) return;
 
     WidgetsBinding.instance!.addPostFrameCallback((_) {
@@ -497,35 +406,18 @@ mixin AnimatedRenderSliverMultiBoxAdaptor
     });
 
     // layout the dragged item
-    _reorderCurrentMainAxisOffset = (_reorderOriginMainAxisOffset +
-            _reorderMainAxisLocalOffset +
-            (constraints.scrollOffset - _reorderOriginScrollOffset))
-        .clamp(0.0, geometry!.maxPaintExtent - _reorderDraggedItemSize);
+    _reorderLayoutData.performLayout(
+        constraints, geometry!.maxPaintExtent, _reorderDraggedItemSize);
 
-    _reorderDraggedRenderBox!
-        .layout(constraints.asBoxConstraints(), parentUsesSize: true);
-
-    parentDataOf(_reorderDraggedRenderBox!)!.layoutOffset =
-        _reorderCurrentMainAxisOffset;
+    layoutWithScrollOffset(
+        _reorderDraggedRenderBox!, _reorderLayoutData.currentMainAxisOffset);
 
     // scroll up/down as needed while dragging
-    var fromOffset = constraints.scrollOffset;
-    var toOffset = fromOffset + constraints.remainingPaintExtent;
-
     var controller = Scrollable.of(childManager)?.widget.controller;
     if (controller != null) {
-      var position = controller.position;
-
-      var delta = 0.0;
-      if (_reorderCurrentMainAxisOffset < fromOffset &&
-          position.extentBefore > 0.0) {
-        delta = -_kReorderingScrollSpeed;
-      } else if (_reorderCurrentMainAxisOffset + _reorderDraggedItemSize >
-              toOffset &&
-          position.extentAfter > 0.0) {
-        delta = _kReorderingScrollSpeed;
-      }
-
+      final position = controller.position;
+      final delta = _reorderLayoutData.scrollDelta(
+          constraints, controller.position, _reorderDraggedItemSize);
       if (delta != 0.0) {
         final value = position.pixels + delta;
         WidgetsBinding.instance!.addPostFrameCallback((_) {
@@ -536,11 +428,11 @@ mixin AnimatedRenderSliverMultiBoxAdaptor
     }
   }
 
-  Rect? _computeItemBox(int buildIndex, bool absolute);
+  Rect? _computeItemBox(int itemIndex, bool absolute, bool avoidMeasuring);
 }
 
-// This class extends the original RenderSliverList to add support for animation and
-// reordering features to a list of variabile-size items.
+/// This class extends the original [RenderSliverList] to add support for animation and
+/// reordering features to a list of variabile-size items.
 class AnimatedRenderSliverList extends RenderSliverList
     with AnimatedRenderSliverMultiBoxAdaptor {
   AnimatedRenderSliverList(AnimatedSliverList widget,
@@ -553,12 +445,12 @@ class AnimatedRenderSliverList extends RenderSliverList
   AnimatedSliverMultiBoxAdaptorElement get childManager =>
       super.childManager as AnimatedSliverMultiBoxAdaptorElement;
 
-  // Measure the size of a bunch of off-list children.
+  // Measures the size of a bunch of off-list children.
   // If the calculated size exceedes the remaining cache extent of the list view,
   // an estimate will be returned.
   @override
   Future<_Measure> measureItems(
-      _Cancelled cancelled, int count, IndexedWidgetBuilder builder) async {
+      _Cancelled? cancelled, int count, IndexedWidgetBuilder builder) async {
     final childConstraints = constraints.asBoxConstraints();
     final maxSize = constraints.remainingCacheExtent;
     assert(count > 0 && maxSize >= 0.0);
@@ -567,10 +459,10 @@ class AnimatedRenderSliverList extends RenderSliverList
     for (; i < count; i++) {
       if (size > maxSize) break;
       await Future.delayed(Duration(milliseconds: 0), () {
-        if (cancelled.value) return _Measure.zero;
+        if (cancelled?.value ?? false) return _Measure.zero;
         size += measureItem(builder.call(childManager, i), childConstraints);
       });
-      if (cancelled.value) return _Measure.zero;
+      if (cancelled?.value ?? false) return _Measure.zero;
     }
     // }
     if (i < count) size *= (count / i);
@@ -580,7 +472,7 @@ class AnimatedRenderSliverList extends RenderSliverList
   @override
   double measureItem(Widget widget, [BoxConstraints? childConstraints]) {
     late double size;
-    childManager.disposableChild(widget, (renderBox) {
+    childManager._disposableElement(widget, (renderBox) {
       renderBox.layout(childConstraints ?? constraints.asBoxConstraints(),
           parentUsesSize: true);
       size = childSize(renderBox);
@@ -601,47 +493,41 @@ class AnimatedRenderSliverList extends RenderSliverList
 
     if (lastIndex == childCount - 1) return trailingScrollOffset;
 
-    var innerSpace = 0.0, trailingSpace = 0.0;
-    var innerCount = 0, trailingCount = 0;
+    var averageTrailingCount = 0.0;
+    var innerCount = 0, trailingCount = 0, leadingCount = 0;
     var buildIndex = 0;
-    var resizingTotalItemSize = 0.0;
-    var resizingTotalItemCount = 0;
+    var averageInnerCount = 0.0;
     for (final interval in _intervalList) {
-      if (interval is _ResizableInterval) {
+      if (interval is _SpaceInterval) {
+        final si = interval as _SpaceInterval;
         if (firstIndex <= buildIndex) {
           if (buildIndex <= lastIndex) {
             // resizing intervals inside the viewport
             innerCount++;
-            innerSpace += interval.currentSize;
+            averageInnerCount += si.averageItemCount;
           } else {
             // resizing intervals outside/after the viewport
             trailingCount++;
-            trailingSpace += interval.currentSize;
+            averageTrailingCount += si.averageItemCount;
           }
+        } else {
+          leadingCount++;
+          averageInnerCount += si.averageItemCount;
         }
-        final aisc = interval.averageItemSizeCount;
-        resizingTotalItemSize += aisc.size;
-        resizingTotalItemCount += aisc.count;
       }
       buildIndex += interval.buildCount;
     }
 
-    var ret = trailingScrollOffset + trailingSpace;
+    var ret = trailingScrollOffset;
 
-    // items outside/after the viewport, excluding resizing intervals
+    // # items outside/after the viewport, excluding resizing intervals
     final remainingCount = childCount - lastIndex - trailingCount - 1;
     if (remainingCount > 0) {
-      // items inside the viewport, excluding resizing intervals
-      var reifiedCount = 1 + lastIndex - firstIndex - innerCount;
-      double averageExtent;
-      // average size of an item (calculated from the items inside the viewport)
-      averageExtent = (trailingScrollOffset -
-              leadingScrollOffset -
-              innerSpace +
-              resizingTotalItemSize) /
-          (reifiedCount + resizingTotalItemCount);
-      ret += averageExtent * remainingCount;
+      final averageExtent = trailingScrollOffset /
+          (1 + lastIndex - innerCount - leadingCount + averageInnerCount);
+      ret += averageExtent * (remainingCount + averageTrailingCount);
     }
+
     return ret;
   }
 
@@ -657,30 +543,41 @@ class AnimatedRenderSliverList extends RenderSliverList
   void performLayout() => doPerformLayout(() => super.performLayout());
 
   @override
-  Rect? _computeItemBox(int buildIndex, bool absolute) {
+  Rect? _computeItemBox(int itemIndex, bool absolute, bool avoidMeasuring) {
+    if (_intervalList.hasPendingUpdates) return null;
+
+    final interval = _intervalList.intervalAtItemIndex(itemIndex);
+    if (interval.interval is! _InListItemInterval &&
+        interval.interval is! _ReadyToChangingInterval) {
+      return null;
+    }
+    var buildIndex = interval.buildIndex + (itemIndex - interval.itemIndex);
+
+    assert(!_intervalList.hasPendingUpdates);
     if (buildIndex < 0 || buildIndex >= _intervalList.buildItemCount) {
       return null;
     }
-    final firstChild = firstChildWithLayout;
     var r = 0.0;
     var s = 0.0;
     if (firstChild == null) {
+      if (avoidMeasuring) return null;
       for (var i = 0; i < buildIndex; i++) {
-        final widget = childManager._build(i, true);
+        final widget = childManager._build(i, measureOnly: true);
         if (widget == null) return null;
         r += measureItem(widget);
       }
-      final widget = childManager._build(buildIndex, true);
+      final widget = childManager._build(buildIndex, measureOnly: true);
       if (widget == null) return null;
       s = measureItem(widget);
     } else {
-      s = childSize(firstChild);
-      final parentData = parentDataOf(firstChild)!;
+      s = childSize(firstChild!);
+      final parentData = _parentDataOf(firstChild)!;
       var i = parentData.index!;
       r = parentData.layoutOffset!;
       if (buildIndex <= i) {
+        if (avoidMeasuring && buildIndex < i) return null;
         while (buildIndex < i) {
-          final widget = childManager._build(--i, true);
+          final widget = childManager._build(--i, measureOnly: true);
           if (widget == null) return null;
           s = measureItem(widget);
           r -= s;
@@ -688,24 +585,25 @@ class AnimatedRenderSliverList extends RenderSliverList
       } else {
         var child = firstChild;
         while (buildIndex > i) {
-          final nextChild = childAfter(child);
+          final nextChild = childAfter(child!);
           if (nextChild == null) break;
-          final parentData = parentDataOf(nextChild)!;
+          final parentData = _parentDataOf(nextChild)!;
           if (parentData.layoutOffset == null) break;
           child = nextChild;
           i = parentData.index!;
           r = parentData.layoutOffset!;
         }
-        s = childSize(child);
+        s = childSize(child!);
         if (buildIndex > i) {
+          if (avoidMeasuring) return null;
           i++;
           r += s;
           while (buildIndex > i) {
-            final widget = childManager._build(i++, true);
+            final widget = childManager._build(i++, measureOnly: true);
             if (widget == null) return null;
             r += measureItem(widget);
           }
-          final widget = childManager._build(buildIndex, true);
+          final widget = childManager._build(buildIndex, measureOnly: true);
           if (widget == null) return null;
           s = measureItem(widget);
         }
@@ -740,16 +638,16 @@ class AnimatedRenderSliverFixedExtentList extends RenderSliverFixedExtentList
       super.childManager as AnimatedSliverMultiBoxAdaptorElement;
 
   @override
-  Future<_Measure> measureItems(_Cancelled cancelled, int count,
+  Future<_Measure> measureItems(_Cancelled? cancelled, int count,
           IndexedWidgetBuilder builder) async =>
-      _Measure(itemExtent * count, false);
+      (itemExtent * count).toExactMeasure();
 
   @override
   double measureItem(Widget widget, [BoxConstraints? childConstraints]) =>
       itemExtent;
 
   @override
-  void _resizingIntervalUpdated(_ResizingInterval interval, double delta) {
+  void _resizingIntervalUpdated(_AnimatedSpaceInterval interval, double delta) {
     super._resizingIntervalUpdated(interval, delta);
     markNeedsLayout();
   }
@@ -764,20 +662,19 @@ class AnimatedRenderSliverFixedExtentList extends RenderSliverFixedExtentList
   ) {
     assert(!_intervalList.hasPendingUpdates);
 
-    if (lastIndex == childCount - 1) {
-      return trailingScrollOffset;
-    }
+    if (lastIndex == childCount - 1) return trailingScrollOffset;
 
     var trailingSpace = 0.0;
     var trailingCount = 0;
     var buildIndex = 0;
     for (final interval in _intervalList) {
-      if (interval is _ResizableInterval) {
+      if (interval is _SpaceInterval) {
+        final si = interval as _SpaceInterval;
         if (firstIndex <= buildIndex) {
           if (buildIndex > lastIndex) {
             // resizing intervals outside/after the viewport
             trailingCount++;
-            trailingSpace += interval.currentSize;
+            trailingSpace += si.currentSize;
           }
         }
       }
@@ -1071,7 +968,7 @@ class AnimatedRenderSliverFixedExtentList extends RenderSliverFixedExtentList
   }
 
   @override
-  Rect? _computeItemBox(int buildIndex, bool absolute) {
+  Rect? _computeItemBox(int buildIndex, bool absolute, bool _) {
     if (buildIndex < 0 || buildIndex >= _intervalList.buildItemCount) {
       return null;
     }
@@ -1085,5 +982,91 @@ class AnimatedRenderSliverFixedExtentList extends RenderSliverFixedExtentList
       case Axis.vertical:
         return Rect.fromLTWH(0, r, constraints.crossAxisExtent, itemExtent);
     }
+  }
+}
+
+class _ReorderLayoutData {
+  var axisLocalOffset = 0.0;
+  var crossAxisLocalOffset = 0.0;
+
+  var originScrollOffset = 0.0;
+
+  var originMainAxisOffset = 0.0;
+  var originCrossAxisOffset = 0.0;
+
+  var currentMainAxisOffset = 0.0;
+  var lastMainAxisOffset = 0.0;
+
+  void init(SliverConstraints constraints, double layoutOffset, double dx,
+      double dy) {
+    update(constraints, dx, dy);
+
+    originCrossAxisOffset = crossAxisLocalOffset;
+
+    originScrollOffset = constraints.scrollOffset;
+
+    currentMainAxisOffset = layoutOffset;
+
+    lastMainAxisOffset =
+        originMainAxisOffset = currentMainAxisOffset - axisLocalOffset;
+  }
+
+  void update(SliverConstraints constraints, double dx, double dy) {
+    switch (applyGrowthDirectionToAxisDirection(
+        constraints.axisDirection, constraints.growthDirection)) {
+      case AxisDirection.up:
+        crossAxisLocalOffset = dx;
+        axisLocalOffset = -dy;
+        break;
+      case AxisDirection.right:
+        axisLocalOffset = dx;
+        crossAxisLocalOffset = dy;
+        break;
+      case AxisDirection.down:
+        crossAxisLocalOffset = dx;
+        axisLocalOffset = dy;
+        break;
+      case AxisDirection.left:
+        axisLocalOffset = -dx;
+        crossAxisLocalOffset = dy;
+        break;
+    }
+  }
+
+  double get crossAxisDelta => crossAxisLocalOffset - originCrossAxisOffset;
+
+  int updateDirection() {
+    final current = currentMainAxisOffset;
+    if ((current - lastMainAxisOffset).abs() > 1E-3) {
+      if (current > lastMainAxisOffset) {
+        lastMainAxisOffset = current;
+        return 1;
+      } else {
+        lastMainAxisOffset = current;
+        return -1;
+      }
+    }
+    return 0;
+  }
+
+  void performLayout(
+      SliverConstraints constraints, double maxPaintExtent, double itemSize) {
+    currentMainAxisOffset = (originMainAxisOffset +
+            axisLocalOffset +
+            (constraints.scrollOffset - originScrollOffset))
+        .clamp(0.0, maxPaintExtent - itemSize);
+  }
+
+  double scrollDelta(
+      SliverConstraints constraints, ScrollPosition position, double itemSize) {
+    var fromOffset = constraints.scrollOffset;
+    var toOffset = fromOffset + constraints.remainingPaintExtent;
+    if (currentMainAxisOffset < fromOffset && position.extentBefore > 0.0) {
+      return -_kReorderingScrollSpeed;
+    } else if (currentMainAxisOffset + itemSize > toOffset &&
+        position.extentAfter > 0.0) {
+      return _kReorderingScrollSpeed;
+    }
+    return 0.0;
   }
 }
