+++ pubspec.yaml
@@ -1,24 +1,23 @@
 name: great_list_view
 description: A Flutter package that enhance the standard list view with implicit animations on changes and so on. See README.md file.
-version: 0.1.1
+version: 0.1.2
 homepage: https://github.com/DavideBelsole/great_list_view.git
 repository: https://github.com/DavideBelsole/great_list_view.git
 
 environment:
-  sdk: '>=2.12.3 <3.0.0'
+  sdk: '>=2.14.0 <3.0.0'
   flutter: ">=2.0.6"
 
 dependencies:
   flutter:
     sdk: flutter
-  diffutil_dart: ^2.0.0+1
-  worker_manager: ^4.0.2
-  collection: ^1.15.0
+  diffutil_dart: ">=2.0.0"
+  worker_manager: ">=4.0.2"
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  pedantic: ^1.11.0
+  pedantic: ^1.11.1
   
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
